test/web-platform-tests/to-upstream/html/browsers/browsing-the-web/navigating-across-documents/fragment-anchors.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Clicking a link triggers navigation</title>
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#javascript-protocol">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<a id="foo" href="#foo">foo</a>
<a id="bar" href="#bar">foo</a>
<script>
"use strict";
setup({ allow_uncaught_exception: false });

async_test(t => {
  const foo = document.getElementById('foo');
  const bar =document.getElementById('bar');
  foo.click();
  t.step_timeout(() => {
    assert_equals(window.location.hash, "#foo", "After clicking link");
    bar.click();
    t.step_timeout(() => {
      assert_equals(window.location.hash, "#bar", "After clicking second link");
      t.done();
    }, 0);
  }, 0);

  // assert_equals(history.length === 1, undefined, "Before");


  // window.location.href = "javascript:window.clicked = true;";

  // assert_equals(window.clicked, undefined, "After");

  // t.step_timeout(() => {
  //   assert_equals(window.clicked, true, "After plus a queued task");
  //   t.done();
  // }, 0);
}, "simple case");
</script>
